split styles into separate stylesheets
