+++ frontend/src/File Manager/Actions/Delete/Delete.action.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import Button from "../../components/Button/Button";
+import Button from "../../../components/Button/Button";
+import "./Delete.action.scss";
 
 const DeleteAction = ({
-  files,
   selectedFile,
   triggerAction,
   handleDelete,
