+++ frontend/src/File Manager/Actions/Rename.action.jsx
@@ -4,6 +4,8 @@ import { IoWarningOutline } from "react-icons/io5";
 import { useDetectOutsideClick } from "../../hooks/useDetectOutsideClick";
 import Modal from "../../components/Modal/Modal";
 import { getFileExtension } from "../../utils/getFileExtension";
+import NameInput from "../../components/Name Input/NameInput";
+import ErrorTooltip from "../../components/Error Tooltip/ErrorTooltip";
 
 const maxNameLength = 220;
 
@@ -142,26 +144,25 @@ const RenameAction = ({
 
   return (
     <>
-      <div className={`rename-file-container ${activeLayout}`}>
-        <textarea
-          ref={outsideClick.ref}
-          className="rename-file"
-          maxLength={maxNameLength}
-          value={renameFile}
-          onChange={(e) => {
-            setRenameFile(e.target.value);
-            setFileRenameError(false);
-          }}
-          onKeyDown={handleValidateFolderRename}
-          onClick={(e) => e.stopPropagation()}
-          {...(activeLayout === "list" && { rows: 1 })}
+      <NameInput
+        nameInputRef={outsideClick.ref}
+        maxLength={maxNameLength}
+        value={renameFile}
+        onChange={(e) => {
+          setRenameFile(e.target.value);
+          setFileRenameError(false);
+        }}
+        onKeyDown={handleValidateFolderRename}
+        onClick={(e) => e.stopPropagation()}
+        {...(activeLayout === "list" && { rows: 1 })}
+      />
+      {fileRenameError && (
+        <ErrorTooltip
+          message={renameErrorMessage}
+          xPlacement={errorXPlacement}
+          yPlacement={errorYPlacement}
         />
-        {fileRenameError && (
-          <p className={`folder-name-error ${errorXPlacement} ${errorYPlacement}`}>
-            {renameErrorMessage}
-          </p>
-        )}
-      </div>
+      )}
 
       <Modal
         heading={"Rename"}
