+++ frontend/src/File Manager/Files/Files.scss
@@ -68,86 +68,6 @@
       }
     }
 
-    .rename-file {
-      position: relative;
-      min-width: 69%;
-      max-width: -webkit-fill-available;
-      text-align: center;
-      border: 4px solid $primary-color;
-      border-radius: 10px;
-      font-size: 15px;
-      font-family: "Nunito Sans", sans-serif;
-      padding: 2px 7px;
-      resize: none;
-      field-sizing: content;
-      overflow-y: hidden;
-
-      &:focus {
-        outline: none;
-      }
-    }
-
-    .folder-name-error {
-      position: absolute;
-      z-index: 1;
-      bottom: -68px;
-      left: 16px;
-      padding: 8px;
-      width: 292px;
-      border-radius: 5px;
-      background-color: #696969;
-      text-align: left;
-      margin: 0;
-      font-size: 0.9em;
-      &::before {
-        content: "";
-        position: absolute;
-        top: -20%;
-        rotate: -45deg;
-        border: 15px solid #696969;
-        border-color: transparent #696969 transparent transparent;
-      }
-    }
-
-    .folder-name-error.right {
-      left: 16px;
-      &::before {
-        left: 11%;
-      }
-    }
-
-    .folder-name-error.left {
-      left: -180px;
-      &::before {
-        left: 76%;
-        transform: rotate(90deg) scaleX(-1);
-      }
-    }
-
-    .folder-name-error.top {
-      bottom: unset !important;
-      top: -68px;
-      &::before {
-        content: none;
-      }
-      &:after {
-        content: "";
-        position: absolute;
-        bottom: -20%;
-        left: 11%;
-        rotate: -45deg;
-        border: 15px solid #696969;
-        border-color: transparent transparent #696969 transparent;
-      }
-    }
-
-    .folder-name-error.top.left {
-      &::after {
-        left: 76%;
-        transform: rotate(90deg) scaleX(-1);
-      }
-    }
-
     .file-name {
       max-width: 115px;
     }
@@ -165,6 +85,37 @@
 
   .file-moving {
     opacity: 0.5;
+  }
+
+  .file-context-menu-list {
+    font-size: 1.1em;
+
+    ul {
+      list-style-type: none;
+      padding-left: 0;
+      margin: 0;
+
+      li {
+        display: flex;
+        gap: 6px;
+        align-items: center;
+        padding: 6px 22px 6px 14px;
+
+        &:hover {
+          cursor: pointer;
+          background-color: rgb(0, 0, 0, 0.04);
+        }
+      }
+
+      li.disable-paste {
+        opacity: 0.5;
+
+        &:hover {
+          cursor: default;
+          background-color: transparent;
+        }
+      }
+    }
   }
 }
