+++ frontend/src/components/Error Tooltip/ErrorTooltip.jsx
@@ -1,0 +1,7 @@
+import "./ErrorTooltip.scss";
+
+const ErrorTooltip = ({ message, xPlacement, yPlacement }) => {
+  return <p className={`error-tooltip ${xPlacement} ${yPlacement}`}>{message}</p>;
+};
+
+export default ErrorTooltip;
