+++ frontend/src/components/Name Input/NameInput.scss
@@ -1,0 +1,20 @@
+    @import "../../styles/variables";
+
+    .rename-file {
+        position: relative;
+        min-width: 69%;
+        max-width: -webkit-fill-available;
+        text-align: center;
+        border: 4px solid $primary-color;
+        border-radius: 10px;
+        font-size: 15px;
+        font-family: "Nunito Sans", sans-serif;
+        padding: 2px 7px;
+        resize: none;
+        field-sizing: content;
+        overflow-y: hidden;
+
+        &:focus {
+            outline: none;
+        }
+    }
